requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
Pillow==8.3.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
opencv-python
matplotlib